% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmcm_stan.R
\name{bmcm_stan}
\alias{bmcm_stan}
\title{Run stan mixture cure model}
\usage{
bmcm_stan(
  input_data,
  event_type = "PFS",
  tx_name = "IPILIMUMAB",
  n_iter = 2000,
  n_warmup = 1000,
  ...
)
}
\arguments{
\item{input_data}{dataframe with all variables}

\item{event_type}{OS or PFS}

\item{tx_name}{Treatment name; IPILIMUMAB, NIVOLUMAB, NIVOLUMAB+IPILIMUMAB}

\item{n_iter}{int}

\item{n_warmup}{int}

\item{...}{Additional arguments}
}
\description{
Run stan mixture cure model
}
