% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist_cure_fractions.R
\name{hist_cure_fractions}
\alias{hist_cure_fractions}
\title{hist_cure_fractions}
\usage{
hist_cure_fractions(stan_out)
}
\arguments{
\item{stan_out}{Nested list}
}
\value{

}
\description{
hist_cure_fractions
}
