% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_S_event_type.R
\name{plot_S_event_type}
\alias{plot_S_event_type}
\title{plot_S_event_type}
\usage{
plot_S_event_type(stan_out)
}
\arguments{
\item{stan_out}{Nested list of stan output}
}
\description{
Plot results of running stan
relative survival mixture cure model.
}
\examples{
load("data/stan_out.RData")

}
