% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstanbmcm-package.R
\docType{package}
\name{rstanbmcm-package}
\alias{rstanbmcm-package}
\alias{rstanbmcm}
\title{The 'rstanbmcm' package.}
\description{
Bayesian mixture cure model with relative survival
}
\references{
Stan Development Team (2020). RStan: the R interface to Stan. R package version 2.21.2. https://mc-stan.org
}
